Updated phase 2 config
Made a month 30 seconds (so each round is 6 minutes). This is just a
guess, but I think the extra minute will make the game more playable.
If it is too easy we can reduce, but I want people to be able to
actually finish the game. I left the tactic cards alone.